+++ conversation.py
@@ -1,0 +1,43 @@
+import asyncio
+from typing import Dict, List
+from pydantic import BaseModel
+from litellm import acompletion
+
+class Message(BaseModel):
+    role: str
+    message: str
+    summary: str = ""
+
+    async def generate_summary(self) -> str:
+        result = await acompletion(
+            model="openai/local",
+            base_url="http://localhost:1234/v1",
+            api_key="dummy",
+            messages=[
+                {"role": "system", "content": "You are a helpful assistant that summarizes messages. Summary must be a single sentence and not longer than 10 words or the original.."},
+                {"role": "user", "content": self.message},
+            ],
+            max_tokens=100,
+        )
+        return result.choices[0].message.content
+
+    def model_post_init(self, __context) -> None: asyncio.create_task(self._update_summary())
+    async def _update_summary(self): self.summary = await self.generate_summary()
+
+class Conversation(BaseModel):
+    messages: List[Message]
+    max_recent: int
+
+    def get_messages(self) -> List[Dict]:
+        return [{
+            "role": msg.role, 
+            "content": msg.message if i < self.max_recent or msg.role.lower() == "system" else msg.summary}
+            for i, msg in enumerate(self.messages[::-1])][::-1]
+
+    def save(self, filename: str):
+        import json
+        with open(filename, "w") as f: json.dump([msg.model_dump() for msg in self.messages], f)
+
+    def load(self, filename: str):
+        import json
+        with open(filename, "r") as f: self.messages = [Message(**msg) for msg in json.load(f)]
