+++ oi.py
@@ -3,89 +3,12 @@ import json
 import asyncio
 import colorama
 
-from cli_utils import Text
-from tools.tools import function_tool, ToolRegistry
 from typing import List, Dict, Optional
 from litellm import acompletion
-from pydantic import BaseModel
-
-class Message(BaseModel):
-    role: str
-    message: str
-    summary: str = ""
-
-    async def generate_summary(self) -> str:
-        result = await acompletion(
-            model="openai/local",
-            base_url="http://localhost:1234/v1",
-            api_key="dummy",
-            messages=[
-                {"role": "system", "content": "You are a helpful assistant that summarizes messages. Summary must be a single sentence and not longer than 10 words or the original.."},
-                {"role": "user", "content": self.message},
-            ],
-            max_tokens=100,
-        )
-        return result.choices[0].message.content
-
-    def model_post_init(self, __context) -> None: asyncio.create_task(self._update_summary())
-    async def _update_summary(self): self.summary = await self.generate_summary()
-
-class Conversation(BaseModel):
-    messages: List[Message]
-    max_recent: int
-
-    def get_messages(self) -> List[Dict]:
-        return [{
-            "role": msg.role, 
-            "content": msg.message if i < self.max_recent or msg.role.lower() == "system" else msg.summary}
-            for i, msg in enumerate(self.messages[::-1])][::-1]
-
-    def save(self, filename: str):
-        with open(filename, "w") as f: json.dump([msg.model_dump() for msg in self.messages], f)
-
-    def load(self, filename: str):
-        with open(filename, "r") as f: self.messages = [Message(**msg) for msg in json.load(f)]
-
-@function_tool
-def unix_bash(command: str) -> Message:
-    """Runs Bash commands on Unix-like systems."""
-    import subprocess
-    result = subprocess.run(command, shell=True, capture_output=True, text=True)
-    output = result.stdout or result.stderr or "Command executed successfully with no output"
-    return Message(
-        role="tool",
-        message=f"{os.getcwd()}$ {command}\n{output}",
-        summary="",
-    )
-
-@function_tool
-def windows_cmd(command: str) -> Message:
-    """Runs Windows CMD commands."""
-    import subprocess
-    result = subprocess.run(command, shell=True, capture_output=True, text=True)
-    output = result.stdout or result.stderr or "Command executed successfully with no output"
-    return Message(
-        role="tool",
-        message=output,
-        summary="",
-    )
-
-@function_tool
-def get_operating_system() -> Message:
-    """Returns the current operating system."""
-    import platform
-    return Message(
-        role="tool",
-        message=f"The OS is {platform.platform(terse=True)}",
-        summary="",
-    )
-
-@function_tool
-def user_input(prompt: Optional[str]) -> Message:
-    if prompt:
-        print(prompt)
-    text = input(Text(text="You: ", color="blue"))
-    return Message(role="user", message=text, summary="")
+from cli_utils import Text
+from tools.tools import function_tool, ToolRegistry
+from tools.terminal import unix_bash, windows_cmd, get_operating_system, user_input
+from conversation import Message, Conversation
 
 class Interpreter:
     def __init__(self, model: str = "openai/local"):
@@ -129,7 +52,7 @@ class Interpreter:
         self.conversation.messages.append(
             Message(role="user", message=initial_text, summary="")
         )
-        
+
         while True:
             await self.respond()
